Move test files to folder
