+++ public/app/plugins/datasource/tempo/SearchTraceQLEditor/SearchField.test.tsx
@@ -6,7 +6,7 @@ import { LanguageProvider } from '@grafana/data';
 import { TraceqlFilter, TraceqlSearchScope } from '../dataquery.gen';
 import { TempoDatasource } from '../datasource';
 import TempoLanguageProvider from '../language_provider';
-import { initTemplateSrv } from '../test_utils';
+import { initTemplateSrv } from '../test/test_utils';
 import { keywordOperators, numberOperators, operators, stringOperators } from '../traceql/traceql';
 
 import SearchField from './SearchField';
